+++ benchmark/benchmark_block_scan.cpp
@@ -1,6 +1,6 @@
 // MIT License
 //
-// Copyright (c) 2020 Advanced Micro Devices, Inc. All rights reserved.
+// Copyright (c) 2020-2022 Advanced Micro Devices, Inc. All rights reserved.
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -30,64 +30,73 @@
 const size_t DEFAULT_N = 1024 * 1024 * 32;
 #endif
 
-template<
-    class Runner,
-    class T,
-    unsigned int BlockSize,
-    unsigned int ItemsPerThread,
-    unsigned int Trials
->
-__global__
-__launch_bounds__(BlockSize)
-void kernel(const T* input, T* output)
+template<hipcub::BlockScanAlgorithm Algorithm,
+         class T,
+         unsigned int BlockSize,
+         unsigned int ItemsPerThread,
+         unsigned int Trials>
+__global__ __launch_bounds__(BlockSize) void inclusive_kernel(const T* input, T* output)
 {
-    Runner::template run<T, BlockSize, ItemsPerThread, Trials>(input, output);
-}
+    const unsigned int i = hipBlockIdx_x * hipBlockDim_x + hipThreadIdx_x;
 
-template<hipcub::BlockScanAlgorithm Algorithm>
-struct inclusive_scan
-{
-    template<
-        class T,
-        unsigned int BlockSize,
-        unsigned int ItemsPerThread,
-        unsigned int Trials
-    >
-    __device__
-    static void run(const T* input, T* output)
+    T values[ItemsPerThread];
+    for(unsigned int k = 0; k < ItemsPerThread; k++)
     {
-        const unsigned int i = hipBlockIdx_x * hipBlockDim_x + hipThreadIdx_x;
-
-        T values[ItemsPerThread];
-        for(unsigned int k = 0; k < ItemsPerThread; k++)
-        {
-            values[k] = input[i * ItemsPerThread + k];
-        }
-
-        using bscan_t = hipcub::BlockScan<T, BlockSize, Algorithm>;
-        __shared__ typename bscan_t::TempStorage storage;
-
-        #pragma nounroll
-        for(unsigned int trial = 0; trial < Trials; trial++)
-        {
-            bscan_t(storage).InclusiveScan(values, values, hipcub::Sum());
-        }
-
-        for(unsigned int k = 0; k < ItemsPerThread; k++)
-        {
-            output[i * ItemsPerThread + k] = values[k];
-        }
+        values[k] = input[i * ItemsPerThread + k];
     }
 
-};
+    using bscan_t = hipcub::BlockScan<T, BlockSize, Algorithm>;
+    __shared__ typename bscan_t::TempStorage storage;
 
-template<
-    class Benchmark,
-    class T,
-    unsigned int BlockSize,
-    unsigned int ItemsPerThread,
-    unsigned int Trials = 100
->
+#pragma nounroll
+    for(unsigned int trial = 0; trial < Trials; trial++)
+    {
+        bscan_t(storage).InclusiveScan(values, values, hipcub::Sum());
+    }
+
+    for(unsigned int k = 0; k < ItemsPerThread; k++)
+    {
+        output[i * ItemsPerThread + k] = values[k];
+    }
+}
+
+template<hipcub::BlockScanAlgorithm Algorithm,
+         class T,
+         unsigned int BlockSize,
+         unsigned int ItemsPerThread,
+         unsigned int Trials>
+__global__
+    __launch_bounds__(BlockSize) void exclusive_kernel(const T* input, T* output, const T init)
+{
+    const unsigned int i = hipBlockIdx_x * hipBlockDim_x + hipThreadIdx_x;
+
+    T values[ItemsPerThread];
+    for(unsigned int k = 0; k < ItemsPerThread; k++)
+    {
+        values[k] = input[i * ItemsPerThread + k];
+    }
+
+    using bscan_t = hipcub::BlockScan<T, BlockSize, Algorithm>;
+    __shared__ typename bscan_t::TempStorage storage;
+
+#pragma nounroll
+    for(unsigned int trial = 0; trial < Trials; trial++)
+    {
+        bscan_t(storage).ExclusiveScan(values, values, init, hipcub::Sum());
+    }
+
+    for(unsigned int k = 0; k < ItemsPerThread; k++)
+    {
+        output[i * ItemsPerThread + k] = values[k];
+    }
+}
+
+template<bool                       Inclusive,
+         hipcub::BlockScanAlgorithm Algorithm,
+         class T,
+         unsigned int BlockSize,
+         unsigned int ItemsPerThread,
+         unsigned int Trials = 100>
 void run_benchmark(benchmark::State& state, hipStream_t stream, size_t N)
 {
     // Make sure size is a multiple of BlockSize
@@ -111,11 +120,29 @@ void run_benchmark(benchmark::State& state, hipStream_t stream, size_t N)
     for (auto _ : state)
     {
         auto start = std::chrono::high_resolution_clock::now();
-        hipLaunchKernelGGL(
-            HIP_KERNEL_NAME(kernel<Benchmark, T, BlockSize, ItemsPerThread, Trials>),
-            dim3(size/items_per_block), dim3(BlockSize), 0, stream,
-            d_input, d_output
-        );
+        if(Inclusive)
+        {
+            hipLaunchKernelGGL(
+                HIP_KERNEL_NAME(inclusive_kernel<Algorithm, T, BlockSize, ItemsPerThread, Trials>),
+                dim3(size / items_per_block),
+                dim3(BlockSize),
+                0,
+                stream,
+                d_input,
+                d_output);
+        }
+        else
+        {
+            hipLaunchKernelGGL(
+                HIP_KERNEL_NAME(exclusive_kernel<Algorithm, T, BlockSize, ItemsPerThread, Trials>),
+                dim3(size / items_per_block),
+                dim3(BlockSize),
+                0,
+                stream,
+                d_input,
+                d_output,
+                input[0]);
+        }
         HIP_CHECK(hipPeekAtLastError());
         HIP_CHECK(hipDeviceSynchronize());
 
@@ -133,45 +160,46 @@ void run_benchmark(benchmark::State& state, hipStream_t stream, size_t N)
 }
 
 // IPT - items per thread
-#define CREATE_BENCHMARK(T, BS, IPT) \
-    benchmark::RegisterBenchmark( \
-        (std::string("block_scan<Datatype:"#T",Block Size:"#BS",Items Per Thread:"#IPT",SubAlgorithm Name:" + algorithm_name + ">.Method Name:") + method_name).c_str(), \
-        &run_benchmark<Benchmark, T, BS, IPT>, \
-        stream, size \
-    )
+#define CREATE_BENCHMARK(T, BS, IPT)                                                          \
+    benchmark::RegisterBenchmark((std::string("block_scan<Datatype:" #T ",Block Size:" #BS    \
+                                              ",Items Per Thread:" #IPT ",SubAlgorithm Name:" \
+                                              + algorithm_name + ">.Method Name:")            \
+                                  + method_name)                                              \
+                                     .c_str(),                                                \
+                                 &run_benchmark<Inclusive, Algorithm, T, BS, IPT>,            \
+                                 stream,                                                      \
+                                 size)
 
-#define BENCHMARK_TYPE(type, block) \
-    CREATE_BENCHMARK(type, block, 1), \
-    CREATE_BENCHMARK(type, block, 2), \
-    CREATE_BENCHMARK(type, block, 3), \
-    CREATE_BENCHMARK(type, block, 4), \
-    CREATE_BENCHMARK(type, block, 8), \
+// clang-format off
+#define BENCHMARK_TYPE(type, block)    \
+    CREATE_BENCHMARK(type, block, 1),  \
+    CREATE_BENCHMARK(type, block, 3),  \
+    CREATE_BENCHMARK(type, block, 4),  \
+    CREATE_BENCHMARK(type, block, 8),  \
     CREATE_BENCHMARK(type, block, 11), \
     CREATE_BENCHMARK(type, block, 16)
+// clang-format on
 
-template<class Benchmark>
+template<bool Inclusive, hipcub::BlockScanAlgorithm Algorithm>
 void add_benchmarks(std::vector<benchmark::internal::Benchmark*>& benchmarks,
-                    const std::string& method_name,
-                    const std::string& algorithm_name,
-                    hipStream_t stream,
-                    size_t size)
+                    const std::string&                            method_name,
+                    const std::string&                            algorithm_name,
+                    hipStream_t                                   stream,
+                    size_t                                        size)
 {
     using custom_float2 = benchmark_utils::custom_type<float, float>;
     using custom_double2 = benchmark_utils::custom_type<double, double>;
 
-    std::vector<benchmark::internal::Benchmark*> new_benchmarks =
-    {
+    std::vector<benchmark::internal::Benchmark*> new_benchmarks = {
         // When block size is less than or equal to warp size
         BENCHMARK_TYPE(int, 64),
         BENCHMARK_TYPE(float, 64),
         BENCHMARK_TYPE(double, 64),
-        BENCHMARK_TYPE(int8_t, 64),
         BENCHMARK_TYPE(uint8_t, 64),
 
         BENCHMARK_TYPE(int, 256),
         BENCHMARK_TYPE(float, 256),
         BENCHMARK_TYPE(double, 256),
-        BENCHMARK_TYPE(int8_t, 256),
         BENCHMARK_TYPE(uint8_t, 256),
 
         CREATE_BENCHMARK(custom_float2, 256, 1),
@@ -209,21 +237,20 @@ int main(int argc, char *argv[])
 
     // Add benchmarks
     std::vector<benchmark::internal::Benchmark*> benchmarks;
-    // inclusive_scan BLOCK_SCAN_RAKING
-    using inclusive_scan_uws_t = inclusive_scan<hipcub::BlockScanAlgorithm::BLOCK_SCAN_RAKING>;
-    add_benchmarks<inclusive_scan_uws_t>(
-        benchmarks, "inclusive_scan", "BLOCK_SCAN_RAKING", stream, size
-    );
-    // inclusive_scan BLOCK_SCAN_RAKING_MEMOIZE
-    using exclusive_scan_uws_t = inclusive_scan<hipcub::BlockScanAlgorithm::BLOCK_SCAN_RAKING_MEMOIZE>;
-    add_benchmarks<exclusive_scan_uws_t>(
-        benchmarks, "inclusive_scan", "BLOCK_SCAN_RAKING_MEMOIZE", stream, size
-    );
-    // inclusive_scan BLOCK_SCAN_WARP_SCANS
-    using inclusive_scan_rts_t = inclusive_scan<hipcub::BlockScanAlgorithm::BLOCK_SCAN_WARP_SCANS>;
-    add_benchmarks<inclusive_scan_rts_t>(
-        benchmarks, "inclusive_scan", "BLOCK_SCAN_WARP_SCANS", stream, size
-    );
+    // clang-format off
+    add_benchmarks<true, hipcub::BlockScanAlgorithm::BLOCK_SCAN_RAKING>(
+        benchmarks, "inclusive_scan", "BLOCK_SCAN_RAKING", stream, size);
+    add_benchmarks<true, hipcub::BlockScanAlgorithm::BLOCK_SCAN_RAKING_MEMOIZE>(
+        benchmarks, "inclusive_scan", "BLOCK_SCAN_RAKING_MEMOIZE", stream, size);
+    add_benchmarks<true, hipcub::BlockScanAlgorithm::BLOCK_SCAN_WARP_SCANS>(
+        benchmarks, "inclusive_scan", "BLOCK_SCAN_WARP_SCANS", stream, size);
+    add_benchmarks<false, hipcub::BlockScanAlgorithm::BLOCK_SCAN_RAKING>(
+        benchmarks, "exclusive_scan", "BLOCK_SCAN_RAKING", stream, size);
+    add_benchmarks<false, hipcub::BlockScanAlgorithm::BLOCK_SCAN_RAKING_MEMOIZE>(
+        benchmarks, "exclusive_scan", "BLOCK_SCAN_RAKING_MEMOIZE", stream, size);
+    add_benchmarks<false, hipcub::BlockScanAlgorithm::BLOCK_SCAN_WARP_SCANS>(
+        benchmarks, "exclusive_scan", "BLOCK_SCAN_WARP_SCANS", stream, size);
+    // clang-format on
 
     // Use manual timing
     for(auto& b : benchmarks)