+++ CHANGELOG.md
@@ -20,7 +20,7 @@ function or by parameters.
 ### Changed
 - Improved build parallelism of the test suite by splitting up large compilation units for `DeviceRadixSort`, 
 `DeviceSegmentedRadixSort` and `DeviceSegmentedSort`.
-- CUB backend references CUB and Thrust version 1.17.2.
+- CUB backend references CUB and Thrust version 1.17.1.
 
 ### Known Issues
 - `BlockRadixRankMatch` is currently broken under the rocPRIM backend.